type

+++ file-content.c
@@ -17,7 +17,7 @@ int main(int argc, char **argv) {
 	char * buf = malloc(BUF_SIZE);
 
 	int fd = open(argv[1], O_RDONLY);
-	u64 offset = strtoull(argv[2]);
+	unsigned long long offset = strtoull(argv[2]);
 
 	while(1) {
 		lseek(fd, offset, SEEK_SET);
